refactor: extract connectors and remove capitalization
