+++ packages/react-celo/src/components/connector-screen.tsx
@@ -1,7 +1,7 @@
 import React, { ReactElement } from 'react';
 
 import cls from '../utils/tailwind';
-import useTheme from '../utils/useTheme';
+import useTheme from '../hooks/use-theme';
 import Button from './button';
 
 interface Footer {
