+++ packages/react-celo/src/components/modal-container.tsx
@@ -4,7 +4,7 @@ import { isMobile } from 'react-device-detect';
 import { SupportedProviders } from '../constants';
 import { Maybe } from '../types';
 import cls from '../utils/tailwind';
-import useTheme from '../utils/useTheme';
+import useTheme from '../hooks/use-theme';
 
 const styles = cls({
   container: ``,
