+++ packages/react-celo/src/components/qrcode.tsx
@@ -9,7 +9,7 @@ import React, { ReactElement, useMemo } from 'react';
 
 import { QRCodeClass } from '../global';
 import cls from '../utils/tailwind';
-import useTheme from '../utils/useTheme';
+import useTheme from '../hooks/use-theme';
 
 // From https://github.com/soldair/node-qrcode#qr-code-capacity
 const qrCodeCapacity: [QRCodeErrorCorrectionLevel, number][] = [
