+++ packages/react-celo/src/connectors/connectors-by-name.ts
@@ -1,14 +1,12 @@
 import { WalletTypes } from '../constants';
 import { Connector, Network } from '../types';
-import {
-  CeloExtensionWalletConnector,
-  InjectedConnector,
-  LedgerConnector,
-  MetaMaskConnector,
-  PrivateKeyConnector,
-  UnauthenticatedConnector,
-  WalletConnectConnector,
-} from './connectors';
+import CeloExtensionWalletConnector from './celo-extension-wallet';
+import InjectedConnector from './injected';
+import LedgerConnector from './ledger';
+import MetaMaskConnector from './metamask';
+import PrivateKeyConnector from './private-key';
+import UnauthenticatedConnector from './unauthenticated';
+import WalletConnectConnector from './wallet-connect';
 
 /**
  * Connectors for each wallet.
