+++ packages/react-celo/src/connectors/index.ts
@@ -1,2 +1,9 @@
-export * from './connectors';
+export { default as CeloExtensionWalletConnector } from './celo-extension-wallet';
+export * from './common';
 export * from './connectors-by-name';
+export { default as InjectedConnector } from './injected';
+export { default as LedgerConnector } from './ledger';
+export { default as MetaMaskConnector } from './metamask';
+export { default as PrivateKeyConnector } from './private-key';
+export { default as UnauthenticatedConnector } from './unauthenticated';
+export { default as WalletConnectConnector } from './wallet-connect';
