+++ packages/react-celo/src/hooks/use-injected-connector.ts
@@ -1,9 +1,9 @@
 import { useCallback, useEffect } from 'react';
 
+import { InjectedConnector, MetaMaskConnector } from '../connectors';
 import { Connector, Dapp, Maybe, Network } from '../types';
 import { useCeloInternal } from '../use-celo';
 import { CeloTokens } from '../utils/metamask';
-import { InjectedConnector, MetaMaskConnector } from './connectors';
 
 export function useInjectedConnector(
   onSubmit: (connector: Connector) => void,
