+++ packages/react-celo/src/hooks/use-is-mounted.ts
