+++ packages/react-celo/src/modals/connect.tsx
@@ -10,6 +10,8 @@ import ReactModal from 'react-modal';
 import ModalContainer from '../components/modal-container';
 import Tray from '../components/tray';
 import { PROVIDERS, SupportedProviders } from '../constants';
+import useProviders, { walletToProvider } from '../hooks/use-providers';
+import useTheme from '../hooks/use-theme';
 import { ConnectorProps, defaultScreens } from '../screens';
 import Placeholder from '../screens/placeholder';
 import { WalletConnect } from '../screens/wallet-connect';
@@ -24,8 +26,6 @@ import { useCeloInternal } from '../use-celo';
 import { hexToRGB } from '../utils/colors';
 import { defaultProviderSort, SortingPredicate } from '../utils/sort';
 import cls from '../utils/tailwind';
-import useProviders, { walletToProvider } from '../utils/useProviders';
-import useTheme from '../utils/useTheme';
 
 export const styles = cls({
   overlay: isMobile
