+++ packages/react-celo/src/screens/metamask.tsx
@@ -4,11 +4,11 @@ import { isMobile } from 'react-device-detect';
 import Button from '../components/button';
 import ConnectorScreen from '../components/connector-screen';
 import Spinner from '../components/spinner';
-import { useInjectedConnector } from '../connectors/useMetaMaskConnector';
 import { PROVIDERS } from '../constants';
+import { useInjectedConnector } from '../hooks/use-injected-connector';
 import { isEthereumFromMetamask } from '../utils/ethereum';
 import cls from '../utils/tailwind';
-import useTheme from '../utils/useTheme';
+import useTheme from '../hooks/use-theme';
 import { ConnectorProps } from '.';
 
 const styles = cls({
