+++ packages/react-celo/src/screens/placeholder.tsx
@@ -3,7 +3,7 @@ import React from 'react';
 import ConnectorScreen from '../components/connector-screen';
 import { useCeloInternal } from '../use-celo';
 import cls from '../utils/tailwind';
-import useTheme from '../utils/useTheme';
+import useTheme from '../hooks/use-theme';
 
 const styles = cls({
   list: `
