+++ packages/react-celo/src/screens/private-key.tsx
@@ -5,7 +5,7 @@ import ConnectorScreen from '../components/connector-screen';
 import { PrivateKeyConnector } from '../connectors';
 import { useCelo } from '../use-celo';
 import cls from '../utils/tailwind';
-import useTheme from '../utils/useTheme';
+import useTheme from '../hooks/use-theme';
 import { ConnectorProps } from '.';
 
 const styles = cls({
