+++ packages/react-celo/src/utils/fetch-wallet-connect-wallets.ts
